requirements.txt
tqdm
numpy
matplotlib
opencv-python
tensorflow-addons
tensorflow-gpu==2.8.0